Refactored test directory to be in line with main
