+++ src/test/requests/AbstractInsightsResponse.test.ts
