+++ src/test/requests/data/insights/SimpleMetric.test.ts
